First port to cython
This cython version is about twice as fast as the original Pyhton version. Installation is more complicated and dependes on the target machine. So it is not clear wheter this version should become the default version.

The model has to be build again, since the model file of the python and cython versions are not compatible.